java-project/hello-spring/src/main/java/choiyh/hellospring/sample/QuizController.java
package choiyh.hellospring.sample;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
public class QuizController {

    @GetMapping("/quiz")
    public ResponseEntity<String> quiz(@RequestParam("code") int code) {
        switch (code) {
            case 1:
                return ResponseEntity.created(null).body("Created!");
            case 2:
                return ResponseEntity.badRequest().body("Bad Request!");
            default:
                return ResponseEntity.ok().body("OK!");
        }
    }

    @PostMapping("/quiz")
    public ResponseEntity<String> quiz2(@RequestBody Code code) {

        switch (code.value()) {
            case 1:
                return ResponseEntity.status(403).body("Forbidden!");
            default:
                return ResponseEntity.ok().body("OK!");
        }

    }

    // record: 데이터 전달을 목적으로 하는 객체를 빠르고 간편하게 만들기 위한 기능.
    // 생성자, 게터, equals(), hashCode(), toString() 메서드 등을 자동으로 생성한다.
    record Code(int value) {}
}
